requirements.txt
tabula-py
pandas
black
jpype1
openpyxl
jupyterlab